now work with kovan; added rpc call for getting block number using http
